requirements.txt
Django==1.7.4
django-activelink==0.4
django-envelope==1.0
django-honeypot==0.4.0
django-widget-tweaks==1.3
easy-thumbnails==2.2
mysql-python==1.2.3
Pillow==2.7.0
six==1.9.0
